Rearranging tests.
