+++ app/beds/bed_crud.py
@@ -27,6 +27,7 @@ class BedCRUD:
         :param session: SQLAlchemy asyncio AsyncSession
         :return: Bed object; beds.bed_models.Bed
         """
+
         db_bed = Bed(**bed.model_dump())
 
         start = time.time()
@@ -53,6 +54,7 @@ class BedCRUD:
         :param bed_id: a Bed unique identifier
         :return: BedRead object; beds.bed_schemas.BedRead
         """
+
         statement = (
             select(Bed)
             .options(
@@ -90,6 +92,7 @@ class BedCRUD:
         :param limit: Number of rows to return
         :return: Sequence[Bed]; beds.bed_models.Bed
         """
+
         statement = (
             select(Bed).where(Bed.garden_id == garden_id).offset(skip).limit(limit)
         )
@@ -120,6 +123,7 @@ class BedCRUD:
         :param bed_update: BedUpdate object beds.bed_schema.BedUpdate
         :return: Bed object; beds.bed_models.Bed
         """
+
         bed: Bed | None = await session.get(Bed, bed_id)
         if bed:
             bed_data = bed_update.model_dump(exclude_unset=True)
@@ -150,6 +154,7 @@ class BedCRUD:
         :param bed_id: Bed UUID
         :return: bool
         """
+
         bed = await session.get(Bed, bed_id)
         if bed:
             start = time.time()
