+++ tests/integration/routes/test_bed_notes/test_bed_notes_create_routes.py
@@ -1,0 +1,81 @@
+import uuid
+
+import pytest
+from httpx import AsyncClient
+
+from app.gardens.garden_models import Garden
+from app.users.user_models import User
+from tests.helpers.test_helpers import get_auth_headers
+
+
+class TestBedNoteCreateRoutes:
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize(
+        "user_name,expected_status",
+        [
+            ("admin", 200),
+            ("moderator", 200),
+            ("authenticated", 200),
+        ],
+    )
+    async def test_create_bed_note(
+        self,
+        client: AsyncClient,
+        default_user: dict[str, User],
+        default_gardens: dict[str, Garden],
+        user_name: str,
+        expected_status: int,
+    ):
+        test_as = default_user.get(user_name, "")
+        username = test_as.username if isinstance(test_as, User) else ""
+        headers = await get_auth_headers(client=client, user_name=username)
+        garden = default_gardens.get(user_name)
+        if isinstance(garden, Garden):
+            bed_id = garden.beds[0].id
+            response = await client.post(
+                url="/api/bed-notes/",
+                json={"note": "this is a bed note", "bed_id": str(bed_id)},
+                headers=headers,
+            )
+            data = response.json()
+            assert response.status_code == expected_status
+            assert data.get("note") == "this is a bed note"
+        else:
+            pytest.fail(f"No garden found for user {user_name}")
+
+    @pytest.mark.asyncio
+    async def test_create_bed_note_unauthenticated(
+        self,
+        client: AsyncClient,
+        default_gardens: dict[str, Garden],
+    ):
+        headers = await get_auth_headers(client=client, user_name="")
+        garden = default_gardens.get("tester")
+        if isinstance(garden, Garden):
+            bed_id = garden.beds[0].id
+            response = await client.post(
+                url="/api/bed-notes/",
+                json={"note": "this is a bed note", "bed_id": str(bed_id)},
+                headers=headers,
+            )
+            assert response.status_code == 401
+        else:
+            pytest.fail("No garden found for user test_user")
+
+    @pytest.mark.asyncio
+    async def test_create_bed_note_bad_id(
+        self,
+        client: AsyncClient,
+        default_user: dict[str, User],
+    ):
+        test_as = default_user.get("admin", "")
+        username = test_as.username if isinstance(test_as, User) else ""
+        headers = await get_auth_headers(client=client, user_name=username)
+        bad_id = uuid.uuid4()
+        response = await client.post(
+            url="/api/bed-notes/",
+            json={"note": "this is the first note", "bed_id": str(bad_id)},
+            headers=headers,
+        )
+        assert response.status_code == 404
