--- tests/integration/routes/test_garden_note_routes.py
@@ -1,391 +1,0 @@
-import uuid
-
-import pytest
-from httpx import AsyncClient
-
-from app.gardens.garden_models import Garden
-from app.users.user_models import User
-from tests.helpers.test_helpers import get_auth_headers
-
-
-class TestGardenNoteRoutes:
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "user_name,expected_status",
-        [
-            ("admin", 200),
-            ("moderator", 200),
-            ("authenticated", 200),
-        ],
-    )
-    async def test_create_garden_note(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-        user_name: str,
-        expected_status: int,
-    ):
-        test_as = default_user.get(user_name, "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        garden = default_gardens.get(user_name)
-        if isinstance(garden, Garden):
-            garden_id = garden.id
-            response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "this is the first note", "garden_id": str(garden_id)},
-                headers=headers,
-            )
-            data = response.json()
-            assert response.status_code == expected_status
-            assert data.get("note") == "this is the first note"
-        else:
-            pytest.fail(f"No garden found for user {user_name}")
-
-    @pytest.mark.asyncio
-    async def test_create_garden_note_unauthenticated(
-        self,
-        client: AsyncClient,
-        default_gardens: dict[str, Garden],
-    ):
-        headers = await get_auth_headers(client=client, user_name="")
-        garden = default_gardens.get("tester")
-        if isinstance(garden, Garden):
-            response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "this is the first note", "garden_id": str(garden.id)},
-                headers=headers,
-            )
-            assert response.status_code == 401
-        else:
-            pytest.fail("No garden found for user test_user")
-
-    @pytest.mark.asyncio
-    async def test_create_garden_note_bad_id(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-    ):
-        test_as = default_user.get("admin", "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        bad_id = uuid.uuid4()
-        response = await client.post(
-            url="/api/garden-notes/",
-            json={"note": "this is the first note", "garden_id": str(bad_id)},
-            headers=headers,
-        )
-        assert response.status_code == 404
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "user_name,expected_status",
-        [
-            ("admin", 200),
-            ("moderator", 200),
-            ("authenticated", 200),
-        ],
-    )
-    async def test_get_garden_note(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-        user_name: str,
-        expected_status: int,
-    ):
-        test_as = default_user.get(user_name, "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        garden = default_gardens.get(user_name)
-        if isinstance(garden, Garden):
-            garden_id = garden.id
-            response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "this is the first note", "garden_id": str(garden_id)},
-                headers=headers,
-            )
-            new_note = response.json()
-            assert response.status_code == 200
-            if response.status_code == 200:
-                get_response = await client.get(
-                    url=f"/api/garden-notes/{new_note.get('id')}",
-                    headers=headers,
-                )
-                assert get_response.status_code == expected_status
-                note = response.json()
-                assert note.get("note") == "this is the first note"
-        else:
-            pytest.fail(f"No garden found for user {user_name}")
-
-    @pytest.mark.asyncio
-    async def test_get_garden_note_unauthenticated(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-    ):
-        headers = await get_auth_headers(client=client, user_name="")
-        garden = default_gardens.get("authenticated")
-        if isinstance(garden, Garden):
-            post_user = default_user.get("authenticated")
-            username = post_user.username if isinstance(post_user, User) else ""
-            post_headers = await get_auth_headers(client=client, user_name=username)
-            garden_id = garden.id if isinstance(garden, Garden) else garden.get("id")
-            response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "this is the first note", "garden_id": str(garden_id)},
-                headers=post_headers,
-            )
-            new_note = response.json()
-            assert response.status_code == 200
-            if response.status_code == 200:
-                get_response = await client.get(
-                    url=f"/api/garden-notes/{new_note.get('id')}",
-                    headers=headers,
-                )
-                assert get_response.status_code == 401
-                note = response.json()
-                assert note.get("note") == "this is the first note"
-        else:
-            pytest.fail("No garden found for user tester")
-
-    @pytest.mark.asyncio
-    async def test_get_garden_note_bad_id(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-    ):
-        test_as = default_user.get("admin", "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        bad_id = uuid.uuid4()
-        response = await client.get(
-            url=f"/api/garden-notes/{bad_id}",
-            headers=headers,
-        )
-        assert response.status_code == 404
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "user_name,expected_status",
-        [
-            ("admin", 200),
-            ("moderator", 200),
-            ("authenticated", 200),
-        ],
-    )
-    async def test_get_garden_notes(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-        user_name: str,
-        expected_status: int,
-    ):
-        test_as = default_user.get(user_name, "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        garden = default_gardens.get(user_name)
-        if isinstance(garden, Garden):
-            response = await client.get(
-                url=f"/api/garden-notes/notes/{garden.id}",
-                headers=headers,
-            )
-            assert response.status_code == expected_status
-            assert isinstance(response.json(), list)
-
-        else:
-            pytest.fail(f"No garden found for user {user_name}")
-
-    @pytest.mark.asyncio
-    async def test_get_garden_notes_unauthenticated(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-    ):
-        headers = await get_auth_headers(client=client, user_name="")
-        garden = default_gardens.get("authenticated")
-        if isinstance(garden, Garden):
-            garden_id = garden.id
-            response = await client.get(
-                url=f"/api/garden-notes/notes/{garden_id}",
-                headers=headers,
-            )
-            assert response.status_code == 401
-        else:
-            pytest.fail("No garden found for user tester")
-
-    @pytest.mark.asyncio
-    async def test_get_garden_notes_bad_id(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-    ):
-        test_as = default_user.get("admin", "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        bad_id = uuid.uuid4()
-        response = await client.get(
-            url=f"/api/garden-notes/notes/{bad_id}",
-            headers=headers,
-        )
-        assert response.status_code == 404
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "user_name,expected_status",
-        [
-            ("admin", 200),
-            ("moderator", 403),
-            ("authenticated", 403),
-        ],
-    )
-    async def test_update_garden_notes(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-        user_name: str,
-        expected_status: int,
-    ):
-        garden = default_gardens.get("tester")
-        test_user_header = await get_auth_headers(
-            client=client, user_name="test_user_user"
-        )
-
-        test_as = default_user.get(user_name, "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        if isinstance(garden, Garden):
-            response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "this is the first note", "garden_id": str(garden.id)},
-                headers=test_user_header,
-            )
-            note = response.json()
-            response = await client.put(
-                url=f"/api/garden-notes/{note.get('id')}",
-                json={
-                    "note": f"Updated by {username}",
-                },
-                headers=headers,
-            )
-            assert response.status_code == expected_status
-            if response.status_code == 200:
-                data = response.json()
-                assert data.get("note") == f"Updated by {username}"
-
-        else:
-            pytest.fail(f"No garden found for user {user_name}")
-
-    @pytest.mark.asyncio
-    async def test_update_garden_notes_unauthenticated(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-    ):
-        garden = default_gardens.get("tester")
-        test_user_header = await get_auth_headers(
-            client=client, user_name="test_user_user"
-        )
-
-        headers = await get_auth_headers(client=client, user_name="")
-        if isinstance(garden, Garden):
-            response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "this is the first note", "garden_id": str(garden.id)},
-                headers=test_user_header,
-            )
-            note = response.json()
-            response = await client.put(
-                url=f"/api/garden-notes/{note.get('id')}",
-                json={
-                    "note": "Updated by unauthenticated",
-                },
-                headers=headers,
-            )
-            assert response.status_code == 401
-
-        else:
-            pytest.fail("No garden found for user tester")
-
-    @pytest.mark.asyncio
-    async def test_update_garden_note_bad_id(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-    ):
-        test_as = default_user.get("admin", "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        bad_id = uuid.uuid4()
-        response = await client.put(
-            url=f"/api/garden-notes/{bad_id}",
-            json={
-                "note": f"Updated by {username}",
-            },
-            headers=headers,
-        )
-        assert response.status_code == 404
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "user_name,expected_status",
-        [
-            ("", 401),
-            ("admin", 200),
-            ("moderator", 403),
-            ("authenticated", 403),
-        ],
-    )
-    async def test_delete_garden_note(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-        default_gardens: dict[str, Garden],
-        user_name: str,
-        expected_status: int,
-    ):
-        create_as = default_user.get("tester", "")
-        username = create_as.username if isinstance(create_as, User) else ""
-        create_headers = await get_auth_headers(client=client, user_name=username)
-        garden = default_gardens.get("tester")
-        if isinstance(garden, Garden):
-            create_response = await client.post(
-                url="/api/garden-notes/",
-                json={"note": "My new garden note", "garden_id": str(garden.id)},
-                headers=create_headers,
-            )
-            note = create_response.json()
-
-            test_as = default_user.get(user_name, "")
-            username = test_as.username if isinstance(test_as, User) else ""
-            headers = await get_auth_headers(client=client, user_name=username)
-
-            delete_note = await client.delete(
-                url=f"/api/garden-notes/{note.get('id')}", headers=headers
-            )
-            assert delete_note.status_code == expected_status
-
-        else:
-            pytest.fail("No garden found for user tester")
-
-    @pytest.mark.asyncio
-    async def test_delete_garden_note_bad_id(
-        self,
-        client: AsyncClient,
-        default_user: dict[str, User],
-    ):
-        test_as = default_user.get("admin", "")
-        username = test_as.username if isinstance(test_as, User) else ""
-        headers = await get_auth_headers(client=client, user_name=username)
-        bad_id = uuid.uuid4()
-        response = await client.delete(
-            url=f"/api/garden-notes/{bad_id}", headers=headers
-        )
-        assert response.status_code == 404
