+++ tests/integration/routes/test_plants/test_plants_delete_routes.py
@@ -1,0 +1,169 @@
+import uuid
+from unittest.mock import patch
+
+import pytest
+from httpx import AsyncClient
+
+from app.ai.models.ai_recommendation_model import AIRecommendations
+from app.gardens.garden_models import Garden
+from app.users.user_models import User
+from tests.helpers.test_helpers import dummy_ai_recommendations, get_auth_headers
+
+
+class TestPlantDeleteRoutes:
+
+    @pytest.mark.asyncio
+    @patch(
+        "app.plants.plant_routes.get_plant_info", return_value=dummy_ai_recommendations
+    )
+    @pytest.mark.parametrize(
+        "user_name,expected_status",
+        [
+            ("", 401),
+            ("admin", 200),
+            ("moderator", 403),
+            ("authenticated", 403),
+        ],
+    )
+    async def test_delete_plant(
+        self,
+        mock_get_plant_info: AIRecommendations,
+        client: AsyncClient,
+        default_user: dict[str, User],
+        default_gardens: dict[str, Garden],
+        user_name: str,
+        expected_status: int,
+    ):
+        test_as = default_user.get("tester", "")
+        username = test_as.username if isinstance(test_as, User) else ""
+        headers = await get_auth_headers(client=client, user_name=username)
+
+        garden = default_gardens.get("tester", "")
+        if isinstance(garden, Garden):
+            bed = garden.beds[0]
+            create_response = await client.post(
+                url="/api/plants/",
+                json={
+                    "species": "tomato",
+                    "variety": "roma",
+                    "bed_id": str(bed.id),
+                },
+                headers=headers,
+            )
+            plant = create_response.json()
+            plant_id = plant.get("id")
+            delete_as = default_user.get(user_name, "")
+            username = delete_as.username if isinstance(delete_as, User) else ""
+            delete_headers = await get_auth_headers(client=client, user_name=username)
+            response = await client.delete(
+                url=f"/api/plants/{plant_id}",
+                headers=delete_headers,
+            )
+            assert response.status_code == expected_status
+        else:
+            pytest.fail(f"No garden found for user {username}")
+
+    @pytest.mark.asyncio
+    @patch(
+        "app.plants.plant_routes.get_plant_info", return_value=dummy_ai_recommendations
+    )
+    @pytest.mark.parametrize(
+        "user_name,expected_status",
+        [
+            ("admin", 200),
+            ("moderator", 200),
+            ("authenticated", 200),
+        ],
+    )
+    async def test_delete_plant_own(
+        self,
+        mock_get_plant_info: AIRecommendations,
+        client: AsyncClient,
+        default_user: dict[str, User],
+        default_gardens: dict[str, Garden],
+        user_name: str,
+        expected_status: int,
+    ):
+        test_as = default_user.get(user_name, "")
+        username = test_as.username if isinstance(test_as, User) else ""
+        headers = await get_auth_headers(client=client, user_name=username)
+
+        garden = default_gardens.get(user_name)
+        if isinstance(garden, Garden):
+            bed = garden.beds[0]
+            create_response = await client.post(
+                url="/api/plants/",
+                json={
+                    "species": "tomato",
+                    "variety": "roma",
+                    "bed_id": str(bed.id),
+                },
+                headers=headers,
+            )
+            plant = create_response.json()
+            plant_id = plant.get("id")
+            response = await client.delete(
+                url=f"/api/plants/{plant_id}",
+                headers=headers,
+            )
+            assert response.status_code == expected_status
+        else:
+            pytest.fail(f"No garden found for user {username}")
+
+    @pytest.mark.asyncio
+    async def test_delete_plant_bad_id(
+        self,
+        client: AsyncClient,
+        default_user: dict[str, User],
+    ):
+        test_as = default_user.get("authenticated", "")
+        username = test_as.username if isinstance(test_as, User) else ""
+        headers = await get_auth_headers(client=client, user_name=username)
+
+        bad_id = uuid.uuid4()
+
+        response = await client.delete(
+            url=f"/api/plants/{bad_id}",
+            headers=headers,
+        )
+        assert response.status_code == 404
+
+    @pytest.mark.asyncio
+    @patch(
+        "app.plants.plant_routes.get_plant_info", return_value=dummy_ai_recommendations
+    )
+    async def test_delete_plant_unauthenticated(
+        self,
+        mock_get_plant_info: AIRecommendations,
+        client: AsyncClient,
+        default_user: dict[str, User],
+        default_gardens: dict[str, Garden],
+    ):
+        test_as = default_user.get("tester", "")
+        username = test_as.username if isinstance(test_as, User) else ""
+        headers = await get_auth_headers(client=client, user_name=username)
+
+        garden = default_gardens.get("tester")
+        if isinstance(garden, Garden):
+            bed = garden.beds[0]
+            create_response = await client.post(
+                url="/api/plants/",
+                json={
+                    "species": "tomato",
+                    "variety": "roma",
+                    "bed_id": str(bed.id),
+                },
+                headers=headers,
+            )
+            plant = create_response.json()
+            plant_id = plant.get("id")
+
+            new_headers = await get_auth_headers(client=client, user_name="")
+            new_response = await client.delete(
+                url=f"/api/plants/{plant_id}",
+                headers=new_headers,
+            )
+
+            assert new_response.status_code == 401
+        else:
+            pytest.fail("No garden found for user tester")
